fix redox styling
