+++ packages/docs/src/components/redoc.js
@@ -1,3 +1,5 @@
+// This component is a hack for Redoc to work with Gatsby
+// We import the missing dependencies for redoc directly on this file
 import { RedocStandalone } from 'redoc';
 import mobx from 'mobx';
 import styled from 'styled-components';
